Revert "Rabina development"
