+++ RadiantTune/Utils/RTAudioPlayer.swift
@@ -52,12 +52,10 @@ extension RTAudioPlayer: STKAudioPlayerDelegate {
     
     func audioPlayer(_ audioPlayer: STKAudioPlayer, stateChanged state: STKAudioPlayerState, previousState: STKAudioPlayerState) {
         playerState = state
-       
         delegate?.playerStateDidChanged(state: state)
     
         if playerState == .playing {
             NotificationCenter.default.post(name: NSNotification.Name(Constants.FavoritesUpdated), object: nil)
-            NotificationCenter.default.post(name: NSNotification.Name(Constants.StationPlaying), object: nil)
         }
         
         if playerState == .stopped {
