+++ RadiantTune/Viewcontrollers/Home/View/RTPlayerWidgetView.swift
@@ -35,26 +35,20 @@ class RTPlayerWidgetView: UIView {
     override init(frame: CGRect) {
         super.init(frame: frame)
         commonInit()
+        setupAnimationView()
      
+        NotificationCenter.default.addObserver(self, selector: #selector(playerStop), name: NSNotification.Name(Constants.StationStopped), object: nil)
     }
     
     required init?(coder: NSCoder) {
         super.init(coder: coder)
         commonInit()
-    }
-    
-    deinit {
-        NotificationCenter.default.removeObserver(self)
-    }
-    
-    @objc private func playerPlaying() {
-        animationView.play()
+        setupAnimationView()
+        NotificationCenter.default.addObserver(self, selector: #selector(playerStop), name: NSNotification.Name(Constants.StationStopped), object: nil)
+ 
         
-        playButton.setImage(UIImage(systemName: "pause.fill"), for: .normal)
-        playButton.isSelected = true
+        
     }
-    
-    
     
     @objc private func playerStop() {
         animationView.stop()
@@ -63,6 +57,9 @@ class RTPlayerWidgetView: UIView {
         playButton.isSelected = false
     }
     
+    private func setupAnimationView() {
+        setupLottieAnimation(animationView, withName: "wave")
+    }
     
 
     func commonInit() {
@@ -84,12 +81,6 @@ class RTPlayerWidgetView: UIView {
         let tapGesture = UITapGestureRecognizer(target: self, action: #selector(iconImageViewAction))
         iconImageView.isUserInteractionEnabled = true
         iconImageView.addGestureRecognizer(tapGesture)
-        
-        NotificationCenter.default.addObserver(self, selector: #selector(playerPlaying), name: NSNotification.Name(Constants.StationPlaying), object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(playerStop), name: NSNotification.Name(Constants.StationStopped), object: nil)
-        
-        setupLottieAnimation(animationView, withName: "wave")
-        
     }
     
     
@@ -129,14 +120,6 @@ class RTPlayerWidgetView: UIView {
             
             // sub Label
             tagsLabel.text = station.country + "|" + station.tags
-            
-            if (playerState == .playing) {
-                playButton.isSelected = true
-                animationView.play()
-            } else {
-                playButton.isSelected = false
-                animationView.stop()
-            }
         }
     }
     
@@ -158,7 +141,6 @@ class RTPlayerWidgetView: UIView {
             } else {
                 // to play
                 RTAudioPlayer.shared.playWith(url: station.url)
-                saveLastPlayedStation(station)
                 animationView.play()
             }
             sender.isSelected = !sender.isSelected
@@ -198,11 +180,4 @@ class RTPlayerWidgetView: UIView {
         }
     }
     
-    private func saveLastPlayedStation(_ station: Station) {
-        RTLastPlayedStationManager.saveLastPlayedStation(station)
-    }
-    
 }
-
-
-
