+++ RadiantTune/Viewcontrollers/Setting/ViewController/RTSettingViewController.storyboard
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="32700.99.1234" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="Y6W-OH-hqX">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="22505" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="Y6W-OH-hqX">
     <device id="retina6_12" orientation="portrait" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="22684"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="22504"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="System colors in document resources" minToolsVersion="11.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
@@ -72,46 +72,16 @@
                                 <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
                                 <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                             </view>
-                            <stackView opaque="NO" contentMode="scaleToFill" alignment="top" spacing="183" translatesAutoresizingMaskIntoConstraints="NO" id="C6S-VM-yfD">
-                                <rect key="frame" x="30" y="242.33333333333337" width="333" height="31"/>
-                                <subviews>
-                                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Auto play" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="092-R5-p9F">
-                                        <rect key="frame" x="0.0" y="0.0" width="101" height="24"/>
-                                        <fontDescription key="fontDescription" type="system" pointSize="20"/>
-                                        <nil key="textColor"/>
-                                        <nil key="highlightedColor"/>
-                                    </label>
-                                    <switch opaque="NO" contentMode="scaleToFill" horizontalHuggingPriority="750" verticalHuggingPriority="750" contentHorizontalAlignment="center" contentVerticalAlignment="center" on="YES" translatesAutoresizingMaskIntoConstraints="NO" id="D1I-IT-YaZ" userLabel="Auto Play Switch">
-                                        <rect key="frame" x="284" y="0.0" width="51" height="31"/>
-                                        <connections>
-                                            <action selector="onAutoPlaySwitchChanged:" destination="Y6W-OH-hqX" eventType="valueChanged" id="0tv-6L-1vn"/>
-                                        </connections>
-                                    </switch>
-                                </subviews>
-                            </stackView>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Enable auto-play to automatically start the radio station when the app is opened." textAlignment="natural" lineBreakMode="wordWrap" numberOfLines="4" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="EuV-4f-Qf7">
-                                <rect key="frame" x="30" y="281.33333333333331" width="280" height="57.333333333333314"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="16"/>
-                                <nil key="textColor"/>
-                                <nil key="highlightedColor"/>
-                            </label>
                         </subviews>
                         <viewLayoutGuide key="safeArea" id="vDu-zF-Fre"/>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                         <constraints>
-                            <constraint firstItem="vDu-zF-Fre" firstAttribute="trailing" secondItem="C6S-VM-yfD" secondAttribute="trailing" constant="30" id="79W-MA-3HA"/>
-                            <constraint firstItem="vDu-zF-Fre" firstAttribute="trailing" secondItem="EuV-4f-Qf7" secondAttribute="trailing" constant="83" id="7OQ-q5-Bvd"/>
                             <constraint firstItem="ZhD-nm-H9L" firstAttribute="leading" secondItem="vDu-zF-Fre" secondAttribute="leading" constant="30" id="8Gg-wi-Zd9"/>
                             <constraint firstAttribute="trailing" secondItem="ZhD-nm-H9L" secondAttribute="trailing" constant="30" id="Doh-kv-K5Y"/>
                             <constraint firstItem="ZhD-nm-H9L" firstAttribute="top" secondItem="vDu-zF-Fre" secondAttribute="top" constant="30" id="KUF-YO-6S7"/>
-                            <constraint firstItem="C6S-VM-yfD" firstAttribute="top" secondItem="ZhD-nm-H9L" secondAttribute="bottom" constant="24" id="ZWV-4d-Wcx"/>
-                            <constraint firstItem="EuV-4f-Qf7" firstAttribute="top" secondItem="C6S-VM-yfD" secondAttribute="bottom" constant="8" id="iHq-uP-u2x"/>
-                            <constraint firstItem="C6S-VM-yfD" firstAttribute="leading" secondItem="vDu-zF-Fre" secondAttribute="leading" constant="30" id="j5d-6F-aao"/>
-                            <constraint firstItem="EuV-4f-Qf7" firstAttribute="leading" secondItem="vDu-zF-Fre" secondAttribute="leading" constant="30" id="uP6-kL-bUc"/>
                         </constraints>
                     </view>
                     <connections>
-                        <outlet property="autoPlaySwitch" destination="D1I-IT-YaZ" id="Abf-AJ-vIU"/>
                         <outlet property="playerWidgetView" destination="BqE-eL-W9q" id="CX3-fS-f1r"/>
                         <outlet property="setTimerBtn" destination="10Z-QM-qqO" id="L4J-AN-Kd4"/>
                         <outlet property="sleepTimerSwitch" destination="khe-r0-lEA" id="fPE-hr-A0B"/>
